requirements.txt
streamlit==1.35.0
google-generativeai==0.5.2
python-dotenv==1.0.1
sentence-transformers==2.6.1
PyMuPDF==1.24.2
tqdm==4.66.4
scikit-learn==1.4.2
